nfa: ensure that start and end states are valid
- Add checks in NFA's constructor to ensure that start and endstates are
valid (members of the automata described by the transition table.

- Adds coverage via two unit tests

Note: the coupling between start_state/end_state and the transition
table is indicative of a problem in the design. It should not be the
case that the start/end states are not members of the automata or that
it is even a possibility. One solution could be to have some sort of
iterator/reference to the table, and to store that as the start/end
nodes. In other words, use a reference to a node in the table instead of
a value. Still, this is a small enough issue that it is not worth
changing, at least for the time being