add: NFA.epsilon_closure() and tests
Calculates the epsilon closure of a state.
In the future it might make sense to cache state's epsilon closures for
performance purposes.

Another thing to figure out moving forward is exactly what a cycle in an
epsilon closure means. Does it mean that each member of the cycle
survives after every iteration of state (once that cycle is reached)? Is
this scenario even plausible (because NFAs can only be constructed by 3
constructors)?